// src/common/common-module/src/enums/pennyDrop.enum.ts
export enum PennyStatus {
  COMPLETED = 'completed',
  CREATED = 'created',
  FAILED = 'failed',
  INCOMPLETE = 'incomplete',
  UNKNOWN = 'unknown',
  REJECTED = 'rejected',
  NAME_MISMATCH = 'name_mismatch',
  INVALID = 'invalid',
}

export enum PennyDropType {
  DECENTRO = 'decentro',
  RAZORPAY = 'razorpay',
}

export enum PennyDropNameMismatch {
  THIRD = '3',
}